+++ src/main/kotlin/com/mojh/cms/security/service/UserDetailsServiceImpl.kt
@@ -3,7 +3,7 @@ package com.mojh.cms.security.service
 import com.mojh.cms.common.exception.CouponApplicationException
 import com.mojh.cms.common.exception.ErrorCode
 import com.mojh.cms.member.repository.MemberRepository
-import com.mojh.cms.security.MemberAdapter
+import com.mojh.cms.security.member.MemberAdapter
 import org.springframework.security.core.userdetails.UserDetails
 import org.springframework.security.core.userdetails.UserDetailsService
 import org.springframework.stereotype.Service
